refactor(CSS): remove old resource file of css cluster
